camera_service/requirements.txt
paho-mqtt
face-recognition
opencv-python
